+++ jwp-basic-gradle/src/main/java/next/controller/LogoutUserController.java
@@ -1,13 +1,14 @@
-package next.web.controller;
+package next.controller;
 
 import core.mvcframework.Controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class HomeController implements Controller {
+public class LogoutUserController implements Controller {
 
     @Override
     public String execute(final HttpServletRequest request, final HttpServletResponse response) {
-        return "home";
+        request.getSession().invalidate();
+        return "redirect:/";
     }
 }
