+++ jwp-basic-gradle/src/main/java/next/controller/ProfileController.java
@@ -1,10 +1,10 @@
-package next.web.controller;
+package next.controller;
 
 import core.mvcframework.Controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import next.web.controller.dto.ProfileUserDto;
-import next.web.service.UserService;
+import next.controller.dto.ProfileUserDto;
+import next.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
