+++ jwp-basic-gradle/src/main/java/next/controller/QuestionFormController.java
@@ -1,13 +1,13 @@
-package next.web.controller;
+package next.controller;
 
 import core.mvcframework.Controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ShowFormController implements Controller {
+public class QuestionFormController implements Controller {
 
     @Override
     public String execute(final HttpServletRequest request, final HttpServletResponse response) {
-        return "qna/show";
+        return "qna/form";
     }
 }
