+++ jwp-basic-gradle/src/main/java/next/controller/ShowFormController.java
@@ -1,13 +1,13 @@
-package next.web.controller;
+package next.controller;
 
 import core.mvcframework.Controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class QuestionFormController implements Controller {
+public class ShowFormController implements Controller {
 
     @Override
     public String execute(final HttpServletRequest request, final HttpServletResponse response) {
-        return "qna/form";
+        return "qna/show";
     }
 }
