+++ jwp-basic-gradle/src/main/java/next/controller/UpdateUserController.java
@@ -1,11 +1,11 @@
-package next.web.controller;
+package next.controller;
 
 import core.mvcframework.Controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import next.model.User;
-import next.web.service.UserService;
+import next.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
